move all blockly logic back to only one class
