+++ blockly/src/utils/HideLanguage.java
@@ -1,0 +1,15 @@
+package utils;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * This annotation is used to mark methods that should not be included in the language server
+ *
+ * @see server.LanguageServer LanguageServer
+ */
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface HideLanguage {}
